pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "google-genai>=1.27.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
